refactor : seperate UI and function
